Moving overload resolution to a dedicated class type
