+++ Sources/SwiftRewriterLib/TypeSystem/DefaultTypeSystem.swift
@@ -591,9 +591,9 @@ public class DefaultTypeSystem: TypeSystem {
                 selector.keywords.count - 1 == invocationTypeHints.count {
                 
                 if let method =
-                    _applyOverloadResolution(methods: methods,
-                                             argumentTypes: invocationTypeHints,
-                                             typeSystem: self) {
+                    overloadResolver()
+                        .applyOverloadResolution(methods: methods,
+                                                 argumentTypes: invocationTypeHints) {
                     
                     return method
                 }
@@ -1227,103 +1227,4 @@ func typeNameIn(swiftType: SwiftType) -> String? {
 
 func typeNameIn(nominalType: NominalSwiftType) -> String {
     return nominalType.typeNameValue
-}
-
-// TODO: Consider moving this to a separate object to better expose it to testing
-// and allow caching and other stateful niceties.
-
-///
-/// - precondition:
-///     for all methods `M` in `methods`,
-///     `M.signature.parameters.count == argumentTypes.count`
-///
-func _applyOverloadResolution(methods: [KnownMethod],
-                              argumentTypes: [SwiftType?],
-                              typeSystem: TypeSystem) -> KnownMethod? {
-    
-    let signatures = methods.map { $0.signature }
-    if let index = _applyOverloadResolution(signatures: signatures,
-                                            argumentTypes: argumentTypes,
-                                            typeSystem: typeSystem) {
-        return methods[index]
-    }
-    
-    return nil
-}
-
-/// Returns a matching resolution by index on a given array of signatures.
-///
-/// - precondition:
-///     for all methods `M` in `methods`,
-///     `M.signature.parameters.count == argumentTypes.count`
-///
-func _applyOverloadResolution(signatures: [FunctionSignature],
-                              argumentTypes: [SwiftType?],
-                              typeSystem: TypeSystem) -> Int? {
-    
-    if signatures.isEmpty {
-        return nil
-    }
-    
-    // All argument types are nil, or no argument types are available: no best
-    // candidate can be decided.
-    if argumentTypes.isEmpty || argumentTypes.allSatisfy({ $0 == nil }) {
-        return 0
-    }
-    
-    // Start with a linear search for the first fully matching method signature
-    let allArgumentsPresent = argumentTypes.allSatisfy { $0 != nil }
-    if allArgumentsPresent {
-        outerLoop:
-        for (i, signature) in signatures.enumerated() {
-            for (argIndex, argumentType) in argumentTypes.enumerated() {
-                guard let argumentType = argumentType else {
-                    break outerLoop
-                }
-                
-                let parameterType = signature.parameters[argIndex].type
-                
-                if !typeSystem.typesMatch(argumentType, parameterType, ignoreNullability: false) {
-                    break
-                }
-                
-                if argIndex == argumentTypes.count - 1 {
-                    // Candidate matches fully
-                    return i
-                }
-            }
-        }
-    }
-    
-    // Do a lookup ignoring type nullability to attempt to find best-matching
-    // candidates, now
-    var candidates = Array(signatures.enumerated())
-    
-    for (argIndex, argumentType) in argumentTypes.enumerated() {
-        guard candidates.count > 1, let argumentType = argumentType else {
-            continue
-        }
-        
-        var doWork = true
-        
-        repeat {
-            doWork = false
-            
-            for (i, signature) in candidates.enumerated() {
-                let parameterType =
-                    signature.element.parameters[argIndex].type
-                
-                if !typeSystem.isType(argumentType.deepUnwrapped,
-                                      assignableTo: parameterType.deepUnwrapped) {
-                    
-                    candidates.remove(at: i)
-                    doWork = true
-                    break
-                }
-            }
-        } while doWork && candidates.count > 1
-    }
-    
-    // Return first candidate found
-    return candidates.first?.offset
 }
